full_stack_todo/requirements.txt
passlib~=1.7
pydantic[email]
python-jose[cryptography]
psycopg~=3.1
python-dotenv~=0.21
python_dateutil == 2.8.2
pandas~=1.5
fastapi~=0.96
python_dateutil >= 2.8.2
python-dotenv~=0.21
pydantic~=1.8
uvicorn~=0.22
passlib~=1.7
pydantic[email]
jinja2
python-multipart